requirements.txt
wheel
setuptools
argparse
pycryptodomex
asyncio
faker